Add KeyConverter abstract class
