+++ src/app/core/services/kafka/converters/key-converter.service.ts
@@ -1,75 +1,30 @@
-import { HttpClient } from '@angular/common/http'
-import { Injectable } from '@angular/core'
-import { KeyExport, SchemaMetadata } from 'src/app/shared/models/kafka'
-import * as AvroSchema from 'avro-js'
+import { Injectable } from '@angular/core';
+import { KeyExport } from 'src/app/shared/models/kafka'
 
-import { LogService } from '../../misc/log.service'
-import { TokenService } from '../../token/token.service'
+@Injectable({
+  providedIn: 'root'
+})
+export abstract class KeyConverterService {
+  /**
+   * Converts a Kafka key to a record format based on the topic schema.
+   * @param kafkaKey The Kafka key to convert.
+   * @param topic The topic name to use for schema lookup.
+   * @returns A promise resolving to the converted record.
+   */
+  abstract convertToRecord(kafkaKey: any, topic: string): Promise<any>
 
-@Injectable()
-export class KeyConverterService {
-  schemas = {}
-  specifications
-  URI_schema: string = '/schema/subjects/'
-  URI_version: string = '/versions/'
-  BASE_URI: string
+  /**
+   * Processes the payload into a KeyExport object.
+   * @param payload The payload to process.
+   * @returns A KeyExport object.
+   */
+  abstract processData(payload: any): KeyExport
 
-  constructor(private logger: LogService, private http: HttpClient, public token: TokenService,
-  ) {
-    this.updateURI()
-  }
-
-  updateURI() {
-    return this.token.getURI().then(uri => (this.BASE_URI = uri))
-  }
-
-  async getSchema(topic) {
-    if (this.schemas[topic]) return Promise.resolve(this.schemas[topic])
-    else {
-      const versionStr = this.URI_version + 'latest'
-      const uri =
-        this.BASE_URI +
-        this.URI_schema +
-        topic +
-        '-key' +
-        versionStr
-      const schema = await this.getLatestKafkaSchemaVersion(uri)
-      this.schemas[topic] = schema
-      return schema
-    }
-  }
-
-  convertToRecord(kafkaKey, topic): any {
-    return this.getSchema(topic).then(keySchemaMetadata => {
-      const key = JSON.parse(keySchemaMetadata.schema)
-      const record = {
-        schema: keySchemaMetadata.id,
-        value: this.convertToAvro(key, kafkaKey)
-      }
-      return record
-    })
-  }
-
-  processData(payload) {
-    const key: KeyExport = {
-      sourceId: payload.sourceId,
-      userId: payload.userId,
-      projectId: payload.projectId
-    }
-
-    return key
-  }
-
-  convertToAvro(schema, value): any {
-    return AvroSchema.parse(schema).clone(value, { wrapUnions: true })
-  }
-
-  getLatestKafkaSchemaVersion(uri): Promise<SchemaMetadata> {
-    return this.http
-      .get<SchemaMetadata>(uri)
-      .toPromise()
-      .catch(e => {
-        throw this.logger.error('Failed to get latest Kafka schema versions', e)
-      })
-  }
+  /**
+   * Converts a value to Avro format using the provided schema.
+   * @param schema The schema to use for conversion.
+   * @param value The value to convert.
+   * @returns The Avro-formatted value.
+   */
+  abstract convertToAvro(schema: any, value: any): any
 }
